+++ JavaScriptEngineSwitcher.Core/JsEngineLoadException.cs
@@ -8,7 +8,7 @@
 	public sealed class JsEngineLoadException : JsException
 	{
 		/// <summary>
-		/// Initializes a new instance of the <see cref="JsEngineLoadException"/> class 
+		/// Initializes a new instance of the <see cref="JsEngineLoadException"/> class
 		/// with a specified error message
 		/// </summary>
 		/// <param name="message">The message that describes the error</param>
@@ -17,7 +17,7 @@
 		{ }
 
 		/// <summary>
-		/// Initializes a new instance of the <see cref="JsEngineLoadException"/> class 
+		/// Initializes a new instance of the <see cref="JsEngineLoadException"/> class
 		/// with a specified error message and a reference to the inner exception that is the cause of this exception
 		/// </summary>
 		/// <param name="message">The error message that explains the reason for the exception</param>
@@ -27,7 +27,7 @@
 		{ }
 
 		/// <summary>
-		/// Initializes a new instance of the <see cref="JsEngineLoadException"/> class 
+		/// Initializes a new instance of the <see cref="JsEngineLoadException"/> class
 		/// with a specified error message and a reference to the inner exception that is the cause of this exception
 		/// </summary>
 		/// <param name="message">The error message that explains the reason for the exception</param>
@@ -38,14 +38,14 @@
 		{ }
 
 		/// <summary>
-		/// Initializes a new instance of the <see cref="JsEngineLoadException"/> class 
+		/// Initializes a new instance of the <see cref="JsEngineLoadException"/> class
 		/// with a specified error message and a reference to the inner exception that is the cause of this exception
 		/// </summary>
 		/// <param name="message">The error message that explains the reason for the exception</param>
 		/// <param name="engineName">Name of JavaScript engine</param>
 		/// <param name="engineVersion">Version of original JavaScript engine</param>
 		/// <param name="innerException">The exception that is the cause of the current exception</param>
-		public JsEngineLoadException(string message, string engineName, string engineVersion, 
+		public JsEngineLoadException(string message, string engineName, string engineVersion,
 			Exception innerException)
 			: base(message, engineName, engineVersion, innerException)
 		{ }