+++ JavaScriptEngineSwitcher.Core/JsException.cs
@@ -27,7 +27,7 @@
 
 
 		/// <summary>
-		/// Initializes a new instance of the <see cref="JsException"/> class 
+		/// Initializes a new instance of the <see cref="JsException"/> class
 		/// with a specified error message
 		/// </summary>
 		/// <param name="message">The message that describes the error</param>
@@ -36,7 +36,7 @@
 		{ }
 
 		/// <summary>
-		/// Initializes a new instance of the <see cref="JsException"/> class 
+		/// Initializes a new instance of the <see cref="JsException"/> class
 		/// with a specified error message and a reference to the inner exception that is the cause of this exception
 		/// </summary>
 		/// <param name="message">The error message that explains the reason for the exception</param>
@@ -46,7 +46,7 @@
 		{ }
 
 		/// <summary>
-		/// Initializes a new instance of the <see cref="JsException"/> class 
+		/// Initializes a new instance of the <see cref="JsException"/> class
 		/// with a specified error message and a reference to the inner exception that is the cause of this exception
 		/// </summary>
 		/// <param name="message">The error message that explains the reason for the exception</param>
@@ -57,7 +57,7 @@
 		{ }
 
 		/// <summary>
-		/// Initializes a new instance of the <see cref="JsException"/> class 
+		/// Initializes a new instance of the <see cref="JsException"/> class
 		/// with a specified error message and a reference to the inner exception that is the cause of this exception
 		/// </summary>
 		/// <param name="message">The error message that explains the reason for the exception</param>