+++ JavaScriptEngineSwitcher.Core/JsRuntimeException.cs
@@ -53,7 +53,7 @@
 		}
 
 		/// <summary>
-		/// Initializes a new instance of the <see cref="JsRuntimeException"/> class 
+		/// Initializes a new instance of the <see cref="JsRuntimeException"/> class
 		/// with a specified error message
 		/// </summary>
 		/// <param name="message">The message that describes the error</param>
@@ -62,7 +62,7 @@
 		{ }
 
 		/// <summary>
-		/// Initializes a new instance of the <see cref="JsRuntimeException"/> class 
+		/// Initializes a new instance of the <see cref="JsRuntimeException"/> class
 		/// with a specified error message and a reference to the inner exception that is the cause of this exception
 		/// </summary>
 		/// <param name="message">The error message that explains the reason for the exception</param>
@@ -72,7 +72,7 @@
 		{ }
 
 		/// <summary>
-		/// Initializes a new instance of the <see cref="JsRuntimeException"/> class 
+		/// Initializes a new instance of the <see cref="JsRuntimeException"/> class
 		/// with a specified error message and a reference to the inner exception that is the cause of this exception
 		/// </summary>
 		/// <param name="message">The error message that explains the reason for the exception</param>
@@ -83,14 +83,14 @@
 		{ }
 
 		/// <summary>
-		/// Initializes a new instance of the <see cref="JsRuntimeException"/> class 
+		/// Initializes a new instance of the <see cref="JsRuntimeException"/> class
 		/// with a specified error message and a reference to the inner exception that is the cause of this exception
 		/// </summary>
 		/// <param name="message">The error message that explains the reason for the exception</param>
 		/// <param name="engineName">Name of JavaScript engine</param>
 		/// <param name="engineVersion">Version of original JavaScript engine</param>
 		/// <param name="innerException">The exception that is the cause of the current exception</param>
-		public JsRuntimeException(string message, string engineName, string engineVersion, 
+		public JsRuntimeException(string message, string engineName, string engineVersion,
 			Exception innerException)
 			: base(message, engineName, engineVersion, innerException)
 		{