+++ JavaScriptEngineSwitcher.Tests/Resources/power.js
@@ -1,6 +1,6 @@
 ﻿function power(x, y) {
 	var result = x;
-	
+
 	for (var i = 2; i <= y; i++) {
 		result *= x;
 	}